+++ src/main/java/gift/api/option/OptionRepository.java
@@ -1,5 +1,6 @@
 package gift.api.option;
 
+import gift.api.option.domain.Option;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
