+++ src/main/java/gift/global/resolver/LoginMember.java
@@ -1,4 +1,4 @@
-package gift.global;
+package gift.global.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
